Adds table for libraries
This commit changes the list of available libraries to a table, similar to the one used for the list of apps in documentation/Apps.md. Soon more libraries for the same language will appear and a table will handle it better. It also sorts the libraries in ascending order (case insensitive) by language.