Reorganization of the Apps.md page (#245)
* Big update / redesign

Also added some apps/tools. Removed an app that didn't had any webpage or source code (please add it when it has).

* Layout fixes / removed an android app without a build

For end user it is very difficult to build a mobile app (Android or iOS). So source-code only apps should first provide at least a build (or better have it on Fdroid or Play Store or iTunes, etc).

* Update Apps.md

* Update Apps.md