Please add my app.
Hey guys,

we had agreed that my app would be listed here if I still followed the
GPLv3 and added enough modifications and features to really distinguish
from Tusky.

As of the newest update, we have an AMOLED mode, the ability to
customize the cards and Stories and a fullscreen card view that fill up
all the device space.

In addition to the modifications that had already been made, I feel
like this is distinguishing enough to publish it as one of the apps.

Thanks.