[Apache] Allow access the streaming API with both http and ws
The previous configuration example allowed the streaming API to be accessed with websocket only. It's because Apache stops at the first matching "ProxyPass" directive which forward to a ws:// URL. This is problematic for apps (such as Mastalab) who calls the streaming API with "normal" HTTP requests only.

To work around that, I use RewriteCond/RewriteRule directives who matches websockets requests to the streaming API. The following ProxyPass directive is then used only if the RewriteRule didn't match the request (=non-ws requests).