Add Liberapay link and revise some sentences (#522)
Much of this was written back when Eugen was more or less It and contains a lot of first-person stuff. I changed a couple of these to make it project-person, but wanted to get some feedback on that first. I can do a more extensive edit on those through the documentations if it's okay to do.