feat: add a section on periodic cleanup tasks (#816)
* feat: add a section on periodic cleanup tasks

I asked in Discord today about which `tootctl` tasks are good to run on a regular basis (as there are a lot of cleanup jobs). It seems that `media remove` and `preview_cards remove` are the standard ones that you want to run on a regular basis (e.g. with cron). So I added this to the docs in its own section, under "Things to do after installing Mastodon."

* fix: fix typo