Clarify axis info on media "focus" parameter
Today I read the media/ api endpoint document and was confused about the axis orientation (IE is y up or down). I initially implemented an endpoint call with the wrong Y orientation.

There is a helpful diagram but I did not see it because I foolishly scrolled down, saw the return codes section, assumed that was the end of the document, and didn't scroll further.

Edit adds a reference from the arguments section at the top to the detailed explanation of "focus" at the bottom of the page, so the next fool doesn't make the mistake I did.