Removed reference to StoryBlok
I think it's still good guidance to point users to static HTML files for link inclusion. That's what did the trick for me, because before that, I had the link inside a StoryBlok content block and verification didn't work. 

But I agree that there's no real need to mention a specific platform.