layout partial - list - sequential links
Added a partial that creates sequential page navigation. At bottom of each page. Thought it would make the site a bit better as when reading sequentially you can click to the next page in the current section. 

Some of the links had to be modified with JavaScript as was not able to achieve sequential weight navigation with Hugo.